+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-04-18 20:55:57
-// Commit: test(auth): implement widget integration tests for auth
+// Updated at: 2026-04-18 21:48:24
+// Commit: feat(auth): implement shimmer loading skeleton in auth
 
 class authComponentState {
-  final String commitRef = '537';
-  final String lastMessage = 'test(auth): implement widget integration tests for auth';
-  final DateTime timestamp = DateTime.parse('2026-04-18T20:55:57+05:00');
+  final String commitRef = '539';
+  final String lastMessage = 'feat(auth): implement shimmer loading skeleton in auth';
+  final DateTime timestamp = DateTime.parse('2026-04-18T21:48:24+05:00');
 }